+++ boilerplate/types/base/taskrunner/tasks/javascript.js
@@ -9,10 +9,11 @@ import webpack from "webpack";
 import webpackStream from "webpack-stream";
 
 import { process } from "../modules/preprocess-utils.js";
+import { reload } from "./livereload.js";
 
 function javascriptBuildTask(callback)
 {
-  return webpackStream(config.vendors.webpack(), webpack)
+  return webpackStream(config.vendors.webpack.getConfig(), webpack)
     .on("error", (err) => tasks.error("javascript", callback, err))
     .pipe(process(argv.env))
     .on("error", (err) => tasks.error("javascript", callback, err))
@@ -34,7 +35,7 @@ function javascriptLintTask(callback)
 
 function javascriptTaskComplete(callback)
 {
-  global.browserSync.reload(); // TODO: avoid using a global variable
+  reload();
   tasks.success("javascript", callback);
 }